requirements.txt
PyGithub
python-dotenv
faiss-cpu==1.13.1
langchain
langchain-openai
GitPython
pydantic
rich
openai
pydantic
pandas
matplotlib
numpy
